+++ areas/storage/src/AzureMcp.Storage/Options/Blob/Batch/BatchSetTierOptions.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace AzureMcp.Storage.Options.Blob.Batch;
+
+public class BatchSetTierOptions : BaseContainerOptions
+{
+    [JsonPropertyName(StorageOptionDefinitions.TierName)]
+    public string? Tier { get; set; }
+
+    [JsonPropertyName(StorageOptionDefinitions.BlobNamesParam)]
+    public string[]? BlobNames { get; set; }
+}